back/src/gateway/pyproject.toml
[tool.poetry]
name = "gateway"
version = "0.1.0"
description = "API Gateway on FastAPI"
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
httpx = "^0.27.0"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
